Fix palette navigation wraparound in rotary encoder ALT
Changed cast from (unsigned) to (int) in changePalette() function
to prevent wraparound when decrementing palette index from 0.

Previously, when decrementing from the first palette (index 0),
the expression (unsigned)(-1) would wrap around to 4294967295,
causing the palette selection to jump to the end of the palette
list instead of staying at the first palette.

This fix ensures palette navigation stays at boundaries, matching
the behavior of effect navigation.

+++ usermods/usermod_v2_rotary_encoder_ui_ALT/usermod_v2_rotary_encoder_ui_ALT.cpp
@@ -892,7 +892,7 @@ void RotaryEncoderUIUsermod::changePalette(bool increase) {
   }
   display->updateRedrawTime();
 #endif
-  effectPaletteIndex = max(min((unsigned)(increase ? effectPaletteIndex+1 : effectPaletteIndex-1), getPaletteCount()+customPalettes.size()-1), 0U);
+  effectPaletteIndex = max(min((int)(increase ? effectPaletteIndex+1 : effectPaletteIndex-1), (int)(getPaletteCount()+customPalettes.size()-1)), 0);
   effectPalette = palettes_alpha_indexes[effectPaletteIndex];
   stateChanged = true;
   if (applyToAll) {
